requirements.txt
pyppeteer==0.2.5
tweepy==3.10.0
pyrebase5==5.0.1
requests==2.25.1